+++ RoME/R/check_G1_G2.r
@@ -59,10 +59,15 @@ ResultDataTC_temp = ResultDataTC[ResultDataTC$Species == G1_G2[spe,],]
 if (numberError ==0) {
     write(paste("No error occurred"), file = Errors, append = TRUE)
   }
-unlink(file.path(tempdir(),"Logfiles"),recursive=T)
-#unlink(file.path(tempdir(),"Graphs"),recursive=T)
-#unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
-
+  if (file.exists(file.path(tempdir(), "Logfiles"))){
+  unlink(file.path(tempdir(),"Logfiles"),recursive=T)
+  }
+  if (file.exists(file.path(tempdir(), "Graphs"))){
+  unlink(file.path(tempdir(),"Graphs"),recursive=T)
+    }
+	if (file.exists(file.path(tempdir(), "files R-Sufi"))){
+  unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
+    }
     return(TRUE)
 
 }
