+++ RoME/R/check_bridles_length.r
@@ -59,10 +59,15 @@ check_bridles_length<-function(DataTA, wd, suffix){
   }	else {
     write(paste("ATTENTION: BRIDLES_LENGTH=100 between 0 and 200 m, BRIDLES_LENGTH=150 over 200 m"), file = Errors, append = TRUE)
   }
+  if (file.exists(file.path(tempdir(), "Logfiles"))){
   unlink(file.path(tempdir(),"Logfiles"),recursive=T)
-  #unlink(file.path(tempdir(),"Graphs"),recursive=T)
-  #unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
-
+  }
+  if (file.exists(file.path(tempdir(), "Graphs"))){
+  unlink(file.path(tempdir(),"Graphs"),recursive=T)
+    }
+	if (file.exists(file.path(tempdir(), "Graphs"))){
+  unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
+    }
    return(TRUE)
 }
