+++ RoME/R/check_date_haul.r
@@ -81,13 +81,18 @@ for (i in 1:nrow(Dataset)){
  if (numberError ==0) {
     write(paste("No error occurred"), file = Errors, append = TRUE)
   }
-unlink(file.path(tempdir(),"Logfiles"),recursive=T)
-
+  if (file.exists(file.path(tempdir(), "Logfiles"))){
+  unlink(file.path(tempdir(),"Logfiles"),recursive=T)
+  }
+  if (file.exists(file.path(tempdir(), "Graphs"))){
+  unlink(file.path(tempdir(),"Graphs"),recursive=T)
+    }
+	if (file.exists(file.path(tempdir(), "files R-Sufi"))){
+  unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
+    }
+	
   if (numberError ==0) {
     return(TRUE)
   } else { return(FALSE) }
-
-#unlink(file.path(tempdir(),"Graphs"),recursive=T)
-#unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
 
 }
