+++ RoME/R/create_haul.r
@@ -70,6 +70,8 @@ create_haul<-function(ResultDataTA,wd,suffix){
     }
   }
   write.table(traits,file=file.path(wd,"files R-Sufi",paste("traits_",ResultData$YEAR[1],"_GSA",ResultData$AREA[1],".csv",sep="")),row.names=FALSE,quote=FALSE,sep=";")
-unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
+if (file.exists(file.path(tempdir(),"files R-Sufi"))){
 
+unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
+}
   }
