+++ RoME/R/scheme_individual_data.r
@@ -77,10 +77,13 @@ write.table(mat_fin, file = file.path(wd,paste("sampling_individual_measures_",R
 # if (numberError ==0) {
 #     write(paste("No error occurred"), file = Errors, append = TRUE)
 #   }
-unlink(file.path(tempdir(),"Logfiles"),recursive=T)
-unlink(file.path(tempdir(),"Graphs"),recursive=T)
-#unlink(file.path(tempdir(),"files R-Sufi"),recursive=T)
-
+  if (file.exists(file.path(tempdir(), "Logfiles"))){
+  unlink(file.path(tempdir(),"Logfiles"),recursive=T)
+  }
+  if (file.exists(file.path(tempdir(), "Graphs"))){
+  unlink(file.path(tempdir(),"Graphs"),recursive=T)
+    }
+	
   #if (numberError ==0) {
     return(TRUE)
   #} else { return(FALSE) }
